+++ labcodes_answer/lab8/kern/syscall/syscall.c
@@ -34,7 +34,7 @@ sys_wait(uint64_t arg[]) {
 static int
 sys_exec(uint64_t arg[]) {
     const char *name = (const char *)arg[0];
-    int argc = (int)arg[1];
+    int64_t argc = (int64_t)arg[1];
     const char **argv = (const char **)arg[2];
     return do_execve(name, argc, argv);
 }
